move tdict out
