+++ src/aioway/_tdicts.py
