+++ src/aioway/_utils/__init__.py
@@ -10,7 +10,6 @@ from .platforms import *
 from .profs import *
 from .renders import *
 from .signs import *
-from .tdicts import *
 from .tensors import *
 from .types import *
 from .typing import *
